chore(dts-core): remove `transform` module
