+++ crates/dts-core/Cargo.toml
@@ -8,9 +8,7 @@ csv = "1.1.6"
 glob = "0.3.0"
 hcl = { path = "../hcl" }
 dts-json = { path = "../dts-json" }
-indexmap = "1.7.0"
 json5 = "0.4.1"
-jsonpath_lib = "0.3.0"
 pathdiff = "0.2.1"
 regex = "1.5.4"
 serde-xml-rs = "0.5.1"
