+++ crates/dts-core/src/parsers/error.rs
@@ -28,8 +28,6 @@ pub enum ParseErrorKind {
     FlatKey,
     /// Error parsing gron.
     Gron,
-    /// Error parsing function signature.
-    FuncSig,
 }
 
 impl Display for ParseErrorKind {
@@ -37,7 +35,6 @@ impl Display for ParseErrorKind {
         match self {
             ParseErrorKind::FlatKey => write!(f, "flat key"),
             ParseErrorKind::Gron => write!(f, "gron"),
-            ParseErrorKind::FuncSig => write!(f, "function signature"),
         }
     }
 }
